\name{Undocumented}

% database queries
\alias{log_count}
\alias{get_ns}
\alias{get_models}
\alias{get_build_log}
\alias{sys_oid2name}
\alias{get_systems}
\alias{get_user_data}
\alias{sys_comb_matrix}
\alias{pred_stat_get}

% database operations
\alias{mongo_del_oid}
\alias{wrote_model_log}
\alias{purge_predictions}
\alias{set_model_status}
\alias{model_db_write}
\alias{pred_stat_write}

% database helpers
\alias{rmongodb.robj2data.frame}
\alias{pred_stat_make}

% Helpers
\alias{bold.allrows}
\alias{is.between}

% Modelling
\alias{gam.mono.con.fun}
\alias{boot2ci}
\alias{boot2ci_PI}
\alias{build_model}
\alias{predict_RT}

% Plotting
\alias{plot_systems}

% Environment to store variables in
\alias{PredRet.env}








%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Undocumented functions
}
\description{
Collection of undocumented functions used internally in the PredRet shiny app.
}
\author{
Jan Stanstrup, \email{stanstrup@gmail.com}.
}

\keyword{internal}
